Funksjon som angir hvilket format en figur skal lagres i. Inneholder også definisjon av fargepaletter.

Formatet er spesifisert gjennom endinga til filnavnet Tillatt: png, pdf, jpg, bmp, tif, svg og ps Velge farger: BlaaRapp, BlaaHNpms287, GronnHNpms624, GronnHNpms342

figtype(
  outfile = "",
  width = 3 * 595,
  height = 3 * 595,
  res = 3 * 72,
  pointsizePDF = 11,
  fargepalett = "BlaaOff"
)

Arguments

outfile

Navn på fil som skal skrives til. Tom tekststreng skriver til skjerm

width

Bredde på output ved gitt oppløsning

height

Høyde på output ved gitt oppløsning

res

Oppløsningen på raster-output

pointsizePDF

Skrifstørrelse i pdf-output

fargepalett

Hvilken fargepalett skal brukes

Value

UtFigFil List med en del figurparametre